% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AutoScript.R
\name{AS.contains.OR}
\alias{AS.contains.OR}
\title{Detect if any substrings are contained in each string vector element
(case-insensitive)}
\usage{
AS.contains.OR(x, match)
}
\arguments{
\item{x}{String vector.}

\item{match}{String vector of substrings to look for within each element of
\code{x}.}
}
\value{
Logical vector indicating whether each element contains a match.
}
\description{
Checks whether each element of a string vector contains any of the specified
substrings using case-insensitive matching.
}
