% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AutoScript.delim.R
\name{AS.delim.replace}
\alias{AS.delim.replace}
\title{Replace delimited string elements (case- and whitespace-insensitive)}
\usage{
AS.delim.replace(x, pattern, replacement, delimiter = ",")
}
\arguments{
\item{x}{String vector containing delimited strings.}

\item{pattern}{String element to match (case- and whitespace-insensitive).}

\item{replacement}{String element replacement (case- and
whitespace-insensitive).}

\item{delimiter}{String delimiter. Default = \code{","}.}
}
\value{
A string vector with replaced elements.
}
\description{
Splits each string in a string vector by a specified delimiter, trims
whitespace, converts to lowercase, replaces matching elements with a new
string element, and rejoins the elements using the same delimiter.
}
\examples{
treatment <- c("capecitabine",
               "LETROZOLE",
               "letrozole, palbociclib",
               "Letrozole,Ribociclib",
               "anastrozole, ribociclib")
treatment <- AS.delim.replace(treatment, "anastrozole", "ai")
treatment <- AS.delim.replace(treatment, "letrozole", "ai")
treatment <- AS.delim.replace(treatment, "palbociclib", "cdk46i")
treatment <- AS.delim.replace(treatment, "ribociclib", "cdk46i")
print(treatment)
}
