% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AutoScript.delim.R
\name{AS.delim.table}
\alias{AS.delim.table}
\title{Frequency table of delimited string elements (case- and
whitespace-insensitive)}
\usage{
AS.delim.table(x, delimiter = ",")
}
\arguments{
\item{x}{String vector containing delimited strings.}

\item{delimiter}{String delimiter. Default = \code{","}.}
}
\value{
A table of element frequencies (case- and whitespace-insensitive).
}
\description{
Splits each string in a string vector by a specified delimiter, trims
whitespace, converts to lowercase, and tabulates the frequency of all unique
elements.
}
\examples{
treatment <- c("capecitabine",
               "LETROZOLE",
               "letrozole, palbociclib",
               "Letrozole,Ribociclib",
               "anastrozole, ribociclib")
print(AS.delim.table(treatment))
}
