% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AutoScript.R
\name{AS.dispersion}
\alias{AS.dispersion}
\title{Dispersion estimate}
\usage{
AS.dispersion(fit)
}
\arguments{
\item{fit}{A fitted \code{glm} object.}
}
\value{
Numeric value of the dispersion estimate.
}
\description{
Computes the ratio of residual deviance to residual degrees of freedom,
used to assess over-dispersion.
}
