% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AutoScript.R
\name{AS.fixdec}
\alias{AS.fixdec}
\title{Fixed decimal places}
\usage{
AS.fixdec(x, digits = 2)
}
\arguments{
\item{x}{A numeric vector.}

\item{digits}{Number of decimal places. Default = \code{2}.}
}
\value{
A string vector representation of \code{x}.
}
\description{
Converts a numeric vector to strings with a fixed number of decimal places,
including trailing zeros. Invalid values are displayed as \code{"N/A"}. Negative
numbers are represented using true minus signs (\verb{\\u2212}) instead of hyphens.
}
