% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AutoScript.format.R
\name{AS.format}
\alias{AS.format}
\title{Format regression results into manuscript-ready tables}
\usage{
AS.format(fit, name = NULL)
}
\arguments{
\item{fit}{A fitted object of class \code{coxph} or \code{glm}.}

\item{name}{Optional string vector of variable names.}
}
\value{
A character matrix representing a manuscript-ready table.
}
\description{
Format regression results into manuscript-ready tables
}
\examples{
library(AutoScript)
library(survival)
data <- survival::veteran
fit <- coxph(Surv(data$time, data$status) ~ as.factor(data$trt))
table2 <- AS.format(fit, name = "Treatment")
print(table2)
}
