% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AutoScript.R
\name{AS.interleave}
\alias{AS.interleave}
\title{Interleave vectors}
\usage{
AS.interleave(a, b)
}
\arguments{
\item{a}{A vector.}

\item{b}{A vector, recycled against \code{a} if shorter.}
}
\value{
A vector of length \code{2 * length(a)} with elements from \code{a} and \code{b}
alternating.
}
\description{
Combines two vectors by alternating their elements, starting with \code{a}.
}
