% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AutoScript.R
\name{AS.rowmin}
\alias{AS.rowmin}
\title{Row-wise minima}
\usage{
AS.rowmin(input)
}
\arguments{
\item{input}{A numeric matrix or data frame.}
}
\value{
A numeric vector of row-wise minima.
}
\description{
Computes the minimum value of each row in a matrix or data frame,
ignoring \code{NA}s. Returns \code{NA} if an entire row is missing.
}
\details{
An example use case in survival analysis is determining the event date for
progression-free survival based on the dates of progression and death.
}
