% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AutoScript.summary.R
\name{AS.summary.count}
\alias{AS.summary.count}
\title{Summarize count variable}
\usage{
AS.summary.count(x, digits.fixed = 2)
}
\arguments{
\item{x}{A numeric vector.}

\item{digits.fixed}{Number of decimal places for the mean count per
observation. Default = \code{2}.}
}
\value{
A string of the form \code{"count (mean count per observation)"}.
}
\description{
Computes a string with the count and mean count per observation.
}
