% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AutoScript.summary.R
\name{AS.summary.HHMM}
\alias{AS.summary.HHMM}
\title{Summarize time-of-day variable in "HH:MM" format}
\usage{
AS.summary.HHMM(x)
}
\arguments{
\item{x}{A string vector of times-of-day in "HH:MM" format.}
}
\value{
A string of the form \verb{"}mean ± SD\verb{"}.
}
\description{
Computes a string with the mean and standard deviation in the form
\verb{"}circular mean ± circular SD\verb{"}.
}
