% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AutoScript.summary.R
\name{AS.summary.log1plinear}
\alias{AS.summary.log1plinear}
\title{Summarize log1p-transformed continuous variable}
\usage{
AS.summary.log1plinear(x, digits.fixed = 2)
}
\arguments{
\item{x}{A numeric vector.}

\item{digits.fixed}{Number of decimal places. Default = \code{2}.}
}
\value{
A string of the form \code{"back-transformed mean ± back-transformed SD"}.
}
\description{
Computes a string with the back-transformed mean and standard deviation from
the log(x + 1) scale in the form \code{"back-transformed mean ± back-transformed SD"}.
}
