% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AutoScript.R
\name{AS.survoutcome}
\alias{AS.survoutcome}
\title{Survival outcomes}
\usage{
AS.survoutcome(startdate, eventdate, reviewdate, divisor = 365.2425/12)
}
\arguments{
\item{startdate}{Vector of start dates.}

\item{eventdate}{Vector of event dates.}

\item{reviewdate}{Vector of last review dates for censored cases.}

\item{divisor}{Unit conversion factor for time. Default = 365.2425/12.}
}
\value{
A numeric matrix with two columns: follow-up time and status (1 = event, 0 = censored).
}
\description{
Computes event times and statuses from start, event, and review dates.
}
