% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AutoScript.R
\name{glm.count}
\alias{glm.count}
\title{Count regression}
\usage{
glm.count(formula, data = NULL)
}
\arguments{
\item{formula}{Formula for the Poisson model, as for the \code{glm()} function.}

\item{data}{Optional data frame containing model variables.}
}
\value{
A fitted object of class \code{glm} ± \code{negbin}.
}
\description{
Fits a Poisson and a negative binomial regression model, returning the model
with the lower Akaike information criterion¹.
}
\details{
Wrapper function for \code{stats::glm()} and \code{MASS:glm.nb()}.
}
\examples{
library(MASS)
data <- MASS::epil
fit <- glm.count(y ~ trt, data = data)
print(class(fit))
print(AS.format(fit, name = c("(Intercept)", "Treatment")))
}
\references{
\enumerate{
\item Akaike, H., 1974. A new look at the statistical model identification.
\emph{IEEE Transactions on Automatic Control}, 19(6), pp. 716–723.
}
}
