% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AutoScript.R
\name{AS.fixdec}
\alias{AS.fixdec}
\title{Fixed decimal places}
\usage{
AS.fixdec(x, digits = 2)
}
\arguments{
\item{x}{A numeric value.}

\item{digits}{Number of decimal places. Default = \code{2}.}
}
\value{
A string representation of \code{x}.
}
\description{
Converts a numeric value to a string with a fixed number of decimal places,
including trailing zeros. Returns \code{"NaN"} if input is \code{NA}.
}
