% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AutoScript.format.R
\name{AS.format}
\alias{AS.format}
\title{Auto-format}
\usage{
AS.format(
  fit,
  name = NULL,
  digits.fixed = 2,
  digits.sig = 2,
  sig.thresh = 0.001
)
}
\arguments{
\item{fit}{A fitted object of class \code{coxph}, \code{glm}, \code{glmerMod},
\code{lm}, or \code{lmerModLmerTest}.}

\item{name}{Optional string vector of variable names.}

\item{digits.fixed}{Number of decimal places for summaries. Default = 2.}

\item{digits.sig}{Number of significant figures for \emph{p}-values. Default = 2.}

\item{sig.thresh}{Threshold below which \emph{p}-values are displayed as
\code{"< threshold"}. Default = 0.001.}
}
\value{
A character matrix representing a manuscript-ready table.
}
\description{
Format regression results into manuscript-ready tables.
}
\details{
\emph{P}-values for \code{lmerModLmerTest} objects are calculated using the
Kenward–Roger method¹.
}
\examples{
library(AutoScript)
library(survival)
data <- survival::veteran
fit <- coxph(Surv(data$time, data$status) ~ as.factor(data$trt))
table2 <- AS.format(fit, name = "Treatment")
print(table2)
}
\references{
\enumerate{
\item Kenward, M.G. and Roger, J.H., 1997. Small sample inference for fixed effects
from restricted maximum likelihood. \emph{Biometrics}, pp. 983–997.
}
}
