% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AutoScript.format.R
\name{AS.format}
\alias{AS.format}
\title{Regression table auto-formatting}
\usage{
AS.format(
  fit,
  name = NULL,
  hetero.name = NULL,
  digits.fixed = 2,
  digits.sig = 2,
  sig.thresh = 0.001
)
}
\arguments{
\item{fit}{A fitted object of class \code{coxme}, \code{coxph}, \code{glm}, \code{glmerMod}, \code{gls},
\code{lm}, or \code{lmerModLmerTest}.}

\item{name}{Optional string vector of coefficient names.}

\item{hetero.name}{Optional string vector of heteroscedasticity parameter
names.}

\item{digits.fixed}{Number of decimal places for summaries. Default = \code{2}.}

\item{digits.sig}{Number of significant figures for \emph{p}-values. Default = \code{2}.}

\item{sig.thresh}{Threshold below which \emph{p}-values are displayed as
\code{"< threshold"} and values larger than \code{1 - threshold} are displayed as
\code{"> 1 - threshold"}. Default = \code{0.001}.}
}
\value{
A character matrix representing a manuscript-ready table.
}
\description{
Format regression results into manuscript-ready tables.
}
\details{
\emph{P}-values for \code{coxph} objects if \code{weights} are used are calculated using the
Huber–White sandwich estimator\ifelse{latex}{\out{$^{1,2}$}}{\ifelse{html}{\out{<sup>1,2</sup>}}{\out{^{1,2}}}} .
\emph{P}-values for \code{lmerModLmerTest} objects are calculated using the Kenward–Roger
method\ifelse{latex}{\out{$^{3}$}}{\ifelse{html}{\out{<sup>3</sup>}}{^3}}.
}
\examples{
# See GitHub README for further examples:
# https://github.com/hongconsulting/AutoScript
library(AutoScript)
library(survival)
data <- survival::veteran
fit <- coxph(Surv(data$time, data$status) ~ as.factor(data$trt))
print(AS.format(fit, name = "Treatment"))
}
\references{
\enumerate{
\item Huber, P., 1967. The behavior of maximum likelihood estimates under
nonstandard conditions. In: \emph{Proceedings of the fifth Berkeley symposium on
mathematical statistics and probability}, 1(1), pp. 221–233. Berkeley:
University of California Press.
\item White, H., 1980. A heteroskedasticity-consistent covariance matrix
estimator and a direct test for heteroskedasticity. \emph{Econometrica: Journal of
the Econometric Society}, pp. 817–838.
\item Kenward, M.G. and Roger, J.H., 1997. Small sample inference for fixed effects
from restricted maximum likelihood. \emph{Biometrics}, pp. 983–997.
}
}
