% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AutoScript.R
\name{AS.signif}
\alias{AS.signif}
\title{Significant figures}
\usage{
AS.signif(x, digits = 2, threshold = 0.001)
}
\arguments{
\item{x}{A numeric value.}

\item{digits}{Number of significant figures. Default = \code{2}.}

\item{threshold}{Lower bound below which values are displayed as
\code{"< threshold"} and values larger than \code{1 - threshold} are displayed as
\code{"> 1 - threshold"}. Default = \code{0.001}.}
}
\value{
A string representation of \code{x}.
}
\description{
Converts a numeric value to a string with the specified number of significant
figures, including trailing zeros. Values smaller than \code{threshold} are
displayed as \code{"< threshold"} and values larger than \code{1 - threshold} are displayed
as \code{"> 1 - threshold"}. Invalid values are displayed as \code{"N/A"}.
}
