% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AutoScript.summary.R
\name{AS.summary.binary}
\alias{AS.summary.binary}
\title{Summarize binary variable}
\usage{
AS.summary.binary(x, digits.fixed = 0)
}
\arguments{
\item{x}{A binary (0 or 1) numeric or logical vector.}

\item{digits.fixed}{Number of decimal places for the percentage. Default = 2.}
}
\value{
A string of the form \code{"count (percent\%)"}.
}
\description{
Computes a string with the count and percentage of \code{1}s.
}
