% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AutoScript.summary.R
\name{AS.summary.KM}
\alias{AS.summary.KM}
\title{Summarize Kaplan–Meier time-to-event outcome}
\usage{
AS.summary.KM(time, status, digits.fixed = 2)
}
\arguments{
\item{time}{Follow-up times.}

\item{status}{Event indicator (\code{1} = event, \code{0} = censored).}

\item{digits.fixed}{Number of decimal places for the summary. Default = \code{2}.}
}
\value{
A string of the form \code{"median (lower to upper)"} or \code{"NR"} if the
median is not reached.
}
\description{
Computes a string with the Kaplan–Meier median survival time and a 95\%
confidence interval using the Brookmeyer–Crowley¹ method with Greenwood's
variance² and a complementary log–log transformation³.
}
\references{
\enumerate{
\item Brookmeyer, R. and Crowley, J., 1982. A confidence interval for the median
survival time. \emph{Biometrics}, pp. 29–41.
\item Greenwood, M., 1926. A report on the natural duration of cancer. In:
\emph{Reports on Public Health and Medical Subjects}, 33, pp. 1–26. London: Her
Majesty’s Stationery Office, Ministry of Health.
\item Klein, J.P., Logan, B., Harhoff, M. and Andersen, P.K., 2007. Analyzing
survival curves at a fixed point in time. \emph{Statistics in Medicine}, 26(24),
pp. 4505–4519.
}
}
