% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AutoScript.summary.R
\name{AS.summary.linear}
\alias{AS.summary.linear}
\title{Summarize continuous variable}
\usage{
AS.summary.linear(x, digits.fixed = 2)
}
\arguments{
\item{x}{A numeric vector.}

\item{digits.fixed}{Number of decimal places. Default = \code{2}.}
}
\value{
A string of the form \code{"mean ± SD"}.
}
\description{
Computes a string with the mean and standard deviation in the form
\code{"mean ± SD"}.
}
