% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AutoScript.summary.R
\name{AS.summary.loglinear}
\alias{AS.summary.loglinear}
\title{Summarize log-transformed continuous variable}
\usage{
AS.summary.loglinear(x, digits.fixed = 2)
}
\arguments{
\item{x}{A numeric vector.}

\item{digits.fixed}{Number of decimal places. Default = 2.}
}
\value{
A string of the form \verb{"}geometric mean ± geometric SD\verb{"}.
}
\description{
Computes a string with the geometric mean and geometric standard deviation in the form
\verb{"}geometric mean ± geometric SD\verb{"}.
}
